[stage-and-actors] rename @animationProperties to animationProperties@.
Especially in the dependencyMapping case of
    ['animationProperties@', '@animationProperties']
where the 'animationProperties@' path equals the widget rootPath, the
dependenyMapping can be reduced to simply 'animationProperties@' as
the rootPath is applied automatically and the string is expanded to
internal = external = 'animationProperties@'.
Also, the case of "@cellContent" was renamed to "cellContent@" to
have a similar naming convention, even though the use case does not
provide another optimization here.