Fix bug: woff2 decompress async version can't load fonts simultaineously.
The issue seems to be that it uses the same memory for consecutive fonts
and we can't read the old buffer after the new buffer has been written.
The async "gap" creates the problem and the asynchronicity is introdued
only to load the wasm module, not for the actual functionality of
decompressing the buffer, that is syncronous anyways.
It is very good, that there's a way to load the sync version of decompress.