[metamodel] fix consecutive metamorphose of unaltered draft...
In the Videoproof Array V2 "Rap Editor" when a none change was
made e.g. when the an item was dragged directly before or after
itself, so that no actual change happened, the metamorphose failed
because in a struct the "dependencies" property was already defined
as not configurable.

A call to metamorphose in a CoherenceFunction triggered the issue.
See the changed files to get an idea.

It used to "burn" the draft, so it couldn't get
metamorphosed again. Because metamorphose either returns
the item itself, but now immutable OR the OLD immutable
that was the base for the draft if there was no change.
BUT when it returned the OLD_STATE it also changed the
draft so that metamorphose couldn't run again. Hence,
other references of the draft couldn't use it's metamorphose
again. This is fixed now, but it seems, maybe, in this
case relying on the PotentialWriteProxy method (or similar)
and having the change propagate back up to the source
could be a good (better) solution as well.