[generic] Add CollapsibleContainer
In contrast to Collapsible this manages its widgets itself, the
widget defintions are injected as CTOR argument.

All widgets are destroyed when collapsed and rebuild when expanded.
Advantage of thuis approach is that the widgets are not updated when
this is collapsed and this don't require any resources.
Disadvantage is that widgets with ephemeral state will lose that state
when collapsed. Persistent state, that goes into metamodel is of course
preserved.

We could maybe create a "safe haven" state in the ComponentWrapper to
tackle this. We could also mark these widgets somehow, so they don't
get destroyed and kept around instead, that would be more like a hybrid
of CollapsibleContainer and Collapsible.