[prosemirror/integration] make proseMirrorDefaultSchema configurable (WIP)
A newer development is that "proseMirrorSchema" can be optional
e.g. for simple, basic editors. In that case, the schema is not
created dynamically from `proseMirrorSchema` and updated on change,
but it is defined statically and not changing anymore.

The `generic-block` got removed as well, it was initially required
but we were not going to use it. the requirement comes from
doc.content = 'block+'  However, without generic-block, all blocks
are in fact, "unknown". It seems like "unknown" was introduced
after "generic-block" as "unkown" is also configured to satisfy
the 'block+' requirement: `content: "inline*", group: "block"`